Add support for `.clang-format-ignore`
A significant refactor (maybe rewrite?) focused around adding support
for a `.clang-format-ignore` file as requested in #55.

Add two new options to the CLI, in the style of eslint:

- `--ext`, which specifies the comma-separated set of file extensions
  to include for formatting. Default: .c,.cc,.cpp,.h
- `--ignore-path`, which specifies the path to a clang-format ignore
  file. Default: .clang-format-ignore.

The `.clang-format-ignore` file is parsed with the `ignore` npm package,
which is compatible with the gitignore spec.

The CLI is rewritten to have the following phases:

- Walk given directory, recursing into all directories, and yielding
  paths to files that end in the given extensions.
- Relativize the yielded paths to the given directory, filter with
  `ignore`, and absolutize the paths.
- Format or check sources.
- Print report.