Remove duplicated code from embedded clang-format internals
Refactor process spawning and stdout capturing setup to a helper
function. This removes noise from `format` and `clangFormatVersion` such
that these functions are easier to read.

This also DRYs up the code.

This was also an opportunity to clean up the event handlers on the child
process. The ordering of `error` and `close` events is well-defined, so
we can avoid tracking error status and instead just remove event
handlers.